+++ src/hubs/BundlerHub/fusebox/ops/Compile.js
@@ -3,11 +3,13 @@ const Config = require('../config/config')
 
 // @NOTE: might have issues if it requires outside of compiling?
 class Compiler {
-  handle({context, bundle, builder, api}) {
+  handle({context, bundle, name, builder, api}) {
     const config = Config.parse({bundle, builder, context})
     const instance = api.FuseBox.init(config)
     const instructions = Instructor.parse({bundle, builder})
-    instance.bundle(instructions)
+    instance.bundle(instructions).then(() => {
+      context.evts.emit(`opted.compiled.${name}`)
+    })
   }
 }