+++ src/hubs/OpsHub/Exec.js
@@ -1,14 +1,15 @@
 class ExecOp {
-  // should it subscribe to the event
-  // or have a .test?
-  handle({context, bundles}) {
-    const {name} = context
-    const path = bundles.getBundles().getOut()
+  handle({context}) {
+    const {name, bundles, builder} = context
+    // bundles.asFull().out.abs
+    const path = bundles.getBundle().outFile()
+
     if (context.polyfills.window) require('jsdom-global')()
     context.debugFor('exec', `🖥  ops.exec for ${name}`, 'blue', path)
 
     const execed = require(path)
-    context.DebugFor('exec', 'exec result:', 'blue', execed)
+    // context.debugFor('exec', 'exec result:', 'blue')
+    console.verbose(execed, {level: 'exec result', color: 'blue'})
     if (typeof execed === 'function') execed()
   }
 }